Moved tests to the appropriate files
